+++ etc/dbus-serialbattery/bms/battery_template.py
