+++ etc/dbus-serialbattery/bms/lifepower.py
