+++ etc/dbus-serialbattery/bms/lltjbd.py
